Catch cookie error when running Artillery test and Provide useful error message
